Adding data analysis libraries
Disclaimer: I am the author of iminuit, resample, jacobi, and numba-stats. 

These are essential libraries for HEP analysis in Python, if you want to work outside of ROOT. iminuit with numba-stats is faster than RooFit. jacobi and resample cover all needs of estimating and propagating statistical uncertainties. iminuit and resample are part of Scikit-HEP.

I also changed the list of nice standard libraries: pathlib is the modern replacement for direct calls to the os module.